feat: create a new file for the ErrorCode enum and add new code 503
