+++ src/enums/errorCodes.ts
@@ -1,0 +1,10 @@
+enum ErrorCode {
+  VALIDATION_ERROR = 400,
+  UNAUTHORIZED_ERROR = 401,
+  FORBIDDEN_ERROR = 403,
+  NOT_FOUND_ERROR = 404,
+  INTERNAL_SERVER_ERROR = 500,
+  SERVICE_UNAVAILABLE_ERROR = 503,
+}
+
+export default ErrorCode;
